reflect namespace hierarchy in tests
